pyproject.toml
[tool.poetry]
name = "truyen-crawler"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12,<3.15"
fastapi = "^0.128.0"
uvicorn = {extras = ["standard"], version = "^0.40.0"}
pydantic-settings = "^2.12.0"
httpx = "^0.28.1"
beautifulsoup4 = "^4.14.3"
lxml = "^6.0.2"
loguru = "^0.7.3"
sqlalchemy = "^2.0.45"
asyncpg = "^0.31.0"
greenlet = "^3.3.0"
sentence-transformers = "^5.2.0"
pgvector = "^0.4.2"
torch = "^2.9.1"
elasticsearch = "<9.0.0"
aiohttp = "^3.13.3"
google-genai = "^0.4.0"


[tool.poetry.group.dev.dependencies]
ruff = "^0.14.13"
black = "^26.1.0"
pytest = "^9.0.2"
pytest-asyncio = "^1.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
